[FIX] error when creating already existing background directory

+++ install.sh
@@ -108,18 +108,20 @@ set_up_conf()
 };
 
 
-mkdir background
-wget --output-document=Firewatch.jpg https://i.imgur.com/b2VoMCYg.jpg
-if [ -f Firewatch.jpg ]; then
-    convert Firewatch.jpg Firewatch.png
-	rm Firewatch.jpg
-    mv Firewatch.png background/Firewatch.png
-fi
-wget --output-document=Firewatch_lock.jpg https://i.imgur.com/MqeXsvkg.jpg
-if [ -f Firewatch_lock.jpg ]; then
-    convert Firewatch_lock.jpg Firewatch_lock.png
-	rm Firewatch_lock.jpg
-    mv Firewatch_lock.png background/Firewatch_lock.png
+if [ ! -d background ]; then
+	mkdir background
+	wget --output-document=Firewatch.jpg https://i.imgur.com/b2VoMCYg.jpg
+	if [ -f Firewatch.jpg ]; then
+		convert Firewatch.jpg Firewatch.png
+		rm Firewatch.jpg
+		mv Firewatch.png background/Firewatch.png
+	fi
+	wget --output-document=Firewatch_lock.jpg https://i.imgur.com/MqeXsvkg.jpg
+	if [ -f Firewatch_lock.jpg ]; then
+		convert Firewatch_lock.jpg Firewatch_lock.png
+		rm Firewatch_lock.jpg
+		mv Firewatch_lock.png background/Firewatch_lock.png
+	fi
 fi
